OmniCipher_MVP/requirements.txt
llama-cpp-python>=0.2.30
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.24.0


pip install -r requirements.txt